The qemu field in the workload spec is required if machine is set
An attempt to tidy up the code in vm.go a bit, although ultimately this will probably
need to be split up.  If you specify a value for machine you must now also specify a
value for the qemu binary to run.  Previously, the code used qemu-system-riscv64 as
the default if the machine field was specified.

Signed-off-by: Mark Ryan <markdryan@rivosinc.com>