ccvm: allow multiple VMs to be created on the same IP
As long as they use different ports for ssh.  This is necessary
on macos as we only seem to be able to use 127.0.0.1.  If the
user hasn't specified a host port for ssh, ccvm will select an
unused port for them, which is a lot easier than trying to figure
out which ports are free on which IP addresses.

Signed-off-by: Mark Ryan <markdryan@rivosinc.com>